fix Platform Functions tests (#522)
combine and refactor tests from two files into one
